+++ opensrp-chw/src/main/java/org/smartregister/chw/fragment/FamilyProfileMemberFragment.java
@@ -2,26 +2,15 @@ package org.smartregister.chw.fragment;
 
 import android.os.Bundle;
 
-import org.smartregister.chw.activity.AboveFiveChildProfileActivity;
-import org.smartregister.chw.activity.AncMemberProfileActivity;
-import org.smartregister.chw.activity.ChildProfileActivity;
-import org.smartregister.chw.activity.FamilyOtherMemberProfileActivity;
 import org.smartregister.chw.activity.FamilyProfileActivity;
-import org.smartregister.chw.activity.PncMemberProfileActivity;
-import org.smartregister.chw.anc.activity.BaseAncMemberProfileActivity;
-import org.smartregister.chw.core.activity.CoreAboveFiveChildProfileActivity;
-import org.smartregister.chw.core.activity.CoreChildProfileActivity;
 import org.smartregister.chw.core.fragment.CoreFamilyProfileMemberFragment;
 import org.smartregister.chw.model.FamilyProfileMemberModel;
-import org.smartregister.chw.pnc.activity.BasePncMemberProfileActivity;
 import org.smartregister.chw.provider.ChwMemberRegisterProvider;
-import org.smartregister.commonregistry.CommonPersonObject;
 import org.smartregister.cursoradapter.RecyclerViewPaginatedAdapter;
 import org.smartregister.family.fragment.BaseFamilyProfileMemberFragment;
 import org.smartregister.family.presenter.BaseFamilyProfileMemberPresenter;
 import org.smartregister.family.util.Constants;
 
-import java.util.HashMap;
 import java.util.Set;
 
 public class FamilyProfileMemberFragment extends CoreFamilyProfileMemberFragment {
@@ -44,55 +33,6 @@ public class FamilyProfileMemberFragment extends CoreFamilyProfileMemberFragment
         this.clientsView.setAdapter(this.clientAdapter);
     }
 
-    @Override
-    protected Class<?> getFamilyOtherMemberProfileActivityClass() {
-        return FamilyOtherMemberProfileActivity.class;
-    }
-
-    @Override
-    protected Class<? extends CoreAboveFiveChildProfileActivity> getAboveFiveChildProfileActivityClass() {
-        return AboveFiveChildProfileActivity.class;
-    }
-
-    @Override
-    protected Class<? extends CoreChildProfileActivity> getChildProfileActivityClass() {
-        return ChildProfileActivity.class;
-    }
-
-    @Override
-    protected Class<? extends BaseAncMemberProfileActivity> getAncMemberProfileActivityClass() {
-        return AncMemberProfileActivity.class;
-    }
-
-    @Override
-    protected Class<? extends BasePncMemberProfileActivity> getPncMemberProfileActivityClass() {
-        return PncMemberProfileActivity.class;
-    }
-
-    @Override
-    protected boolean isAncMember(String baseEntityId) {
-        return getFamilyProfileActivity().getFamilyProfilePresenter().isAncMember(baseEntityId);
-    }
-
-    @Override
-    protected HashMap<String, String> getAncFamilyHeadNameAndPhone(String baseEntityId) {
-        return getFamilyProfileActivity().getFamilyProfilePresenter().getAncFamilyHeadNameAndPhone(baseEntityId);
-    }
-
-    @Override
-    protected CommonPersonObject getAncCommonPersonObject(String baseEntityId) {
-        return getFamilyProfileActivity().getFamilyProfilePresenter().getAncCommonPersonObject(baseEntityId);
-    }
-
-    @Override
-    protected boolean isPncMember(String baseEntityId) {
-        return getFamilyProfileActivity().getFamilyProfilePresenter().isPncMember(baseEntityId);
-    }
-
-    @Override
-    protected CommonPersonObject getPncCommonPersonObject(String baseEntityId) {
-        return getFamilyProfileActivity().getFamilyProfilePresenter().getPncCommonPersonObject(baseEntityId);
-    }
 
     @Override
     protected void initializePresenter() {
@@ -100,6 +40,20 @@ public class FamilyProfileMemberFragment extends CoreFamilyProfileMemberFragment
         String familyHead = getArguments().getString(Constants.INTENT_KEY.FAMILY_HEAD);
         String primaryCareGiver = getArguments().getString(Constants.INTENT_KEY.PRIMARY_CAREGIVER);
         presenter = new BaseFamilyProfileMemberPresenter(this, new FamilyProfileMemberModel(), null, familyBaseEntityId, familyHead, primaryCareGiver);
+    }
+
+    @Override
+    protected void onViewClicked(android.view.View view) {
+        super.onViewClicked(view);
+        int i = view.getId();
+        if (i == org.smartregister.chw.core.R.id.patient_column) {
+            if (view.getTag() != null && view.getTag(org.smartregister.family.R.id.VIEW_ID) == CLICK_VIEW_NORMAL) {
+                getFamilyProfileActivity().goToProfileActivity(view, getArguments());
+            }
+        } else if (i == org.smartregister.chw.core.R.id.next_arrow && view.getTag() != null &&
+                view.getTag(org.smartregister.family.R.id.VIEW_ID) == CLICK_VIEW_NEXT_ARROW) {
+            getFamilyProfileActivity().goToProfileActivity(view, getArguments());
+        }
     }
 
     private FamilyProfileActivity getFamilyProfileActivity() {
